backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
vanna==0.7.7
pymysql==1.1.1
chromadb==0.5.11
pandas==2.2.3
python-dotenv==1.0.1
openai==1.55.3
httpx==0.27.2
sqlalchemy==2.0.32
ollama
python-multipart